requirements.txt
docker-py
flake8
molecule
yamllint
pyyaml
